Updated: Addbox, Table Size and vector2i to float
fixed using wrong rectangle for the add box.
Added size calculation for the Table class
modified test to see Size.
Changed variables to use floats (vector2i to vector2f)
fixed add box being in the wrong position with the wrong size

+++ PersonalFinances/dropdown.h
@@ -31,11 +31,11 @@ public:
 		}
 	}
 
-	Dropdown(const sf::Vector2i& position, int fields, const std::string * const titleStrings);
-	Dropdown(const sf::Vector2i& position, int fields, const std::string * const titleStrings, const std::vector <std::string*>& elements);
+	Dropdown(const sf::Vector2f& position, int fields, const std::string * const titleStrings);
+	Dropdown(const sf::Vector2f& position, int fields, const std::string * const titleStrings, const std::vector <std::string*>& elements);
 
-	Dropdown(const sf::Vector2i& position, int fields, const std::string * const titleStrings, const float innerMargin, const float outerLineThickness, const float mainboxOuterLineThickness);
-	Dropdown(const sf::Vector2i& position, int fields, const std::string * const titleStrings, const std::vector <std::string*>& elements, const float innerMargin, const float outerLineThickness, const float mainboxOuterLineThickness);
+	Dropdown(const sf::Vector2f& position, int fields, const std::string * const titleStrings, const float innerMargin, const float outerLineThickness, const float mainboxOuterLineThickness);
+	Dropdown(const sf::Vector2f& position, int fields, const std::string * const titleStrings, const std::vector <std::string*>& elements, const float innerMargin, const float outerLineThickness, const float mainboxOuterLineThickness);
 	~Dropdown();
 
 	void render(sf::RenderWindow& window) const;
